+++ modules/core/artfish_fishing_unit_ui.R
@@ -4,17 +4,16 @@ artfish_fishing_unit_ui <- function(id){
   ns <- NS(id)
   
   bs4Dash::tabItem(tabName = "artfish_fishing_unit",
-          fluidRow(
-            tags$h2(i18n("ARTFISH_FISHING_UNIT_TITLE"))
-          ),
-          uiOutput(ns("no_release")),
-          fluidRow(
-            column(6,uiOutput(ns("selectors"))),
-            column(3,uiOutput(ns("value"))),
-            column(3,plotlyOutput(ns("accuracy"),height=200))
-          ),
-          uiOutput(ns("bg_results")),
-          uiOutput(ns("sp_results"))
+                   fluidRow(
+                     div(
+                       width = 12, style = "margin:12px;",
+                       tags$h2(i18n("ARTFISH_FISHING_UNIT_TITLE")),tags$h3(class = "text-muted", i18n("ARTFISH_FISHING_UNIT_SUBTITLE"))
+                     )
+                   ),
+                   uiOutput(ns("no_release")),
+                   uiOutput(ns("filter_selectors")),
+                   uiOutput(ns("indicators")),
+                   uiOutput(ns("results"))
           
   )